поправлены замечания
